app.js fixed & routes were created
