+++ client/src/App.js
@@ -1,50 +1,29 @@
-import React ,{useRef}from 'react';
-import Jumbotron from './components/jumbotron';
-
-
-
+import React, { useRef } from 'react';
+import Jumbotron from './components/Jumbotron';
+import { BrowserRouter as Router, Route } from "react-router-dom";
+//import Home from "./pages/Home";
+import Search from "./pages/Search";
+import Signup from "./pages/Signup";
+import Profile from "./pages/Profile";
+//import footer from "./components/footer";
+//import navbar from "./components/navbar";
 
 
 //import './App.css';
 
 function App() {
-  const userRef = useRef()
-  const passRef = useRef()
-  const doIt = ()=>{
-    console.log("we did it") 
-    loginUser(userRef.current.value, passRef.current.value)
-  }
-  function loginUser(email, password) {
-    fetch("/api/login", {
-      email: email,
-      password: password
-    })
-      .then((parameter) => {
-        
-    debugger
-        window.location.replace("/members");
-        // If there's an error, log the error
-      })
-      .catch(err => {
-        console.log(err);
-      });
-  }
-
   return (
-
-
-    <div className="app">
-      <div>
+    <Router>
+      {/* <Navbar /> */}
       <Jumbotron />
-        <h1>Register</h1>
-        <input placeholder="username" ref={userRef}></input>
-        <input placeholder="password" ref={passRef}></input>
-        <button onClick={()=>doIt()}>submit</button>
-        
-      </div>
-    </div>
+      {/* <Route exact path="/" component={Home} /> */}
+      <Route exact path="/Search" component={Search} />
+      <Route exact path="/" component={Signup} />
+      <Route exact path="/Profile" component={Profile} />
+      {/* <Footer /> */}
+    </Router>
+  )
 
-  );
 }
 
 export default App;
