+++ config/config.json
@@ -1,7 +1,7 @@
 {
   "development": {
     "username": "root",
-    "password": "root",
+    "password": "Ilhanirem51",
     "database": "bingewatcher_db",
     "host": "127.0.0.1",
     "dialect": "mysql"
